+++ lib/rspec/core/reporter.rb
@@ -19,6 +19,8 @@ module RSpec::Core
       @pending_examples = []
       @duration = @start = @load_time = nil
       @non_example_exception_count = 0
+      @setup_default = lambda {}
+      @setup = false
     end
 
     # @private
@@ -50,6 +52,13 @@ module RSpec::Core
         @listeners[notification.to_sym] << listener
       end
       true
+    end
+
+    # @private
+    def prepare_default(loader, output_stream, deprecation_stream)
+      @setup_default = lambda do
+        loader.setup_default output_stream, deprecation_stream
+      end
     end
 
     # @private
@@ -200,6 +209,7 @@ module RSpec::Core
 
     # @private
     def notify(event, notification)
+      ensure_listeners_ready
       registered_listeners(event).each do |formatter|
         formatter.__send__(event, notification)
       end
@@ -224,6 +234,13 @@ module RSpec::Core
     end
 
   private
+
+    def ensure_listeners_ready
+      return if @setup
+
+      @setup_default.call
+      @setup = true
+    end
 
     def close
       notify :close, Notifications::NullNotification