+++ spec/rspec/core/configuration_spec.rb
@@ -1459,6 +1459,53 @@ module RSpec::Core
       end
     end
 
+    describe '#reporter' do
+      before do
+        config.output_stream = StringIO.new
+        config.deprecation_stream = StringIO.new
+      end
+
+      it 'does not immediately trigger formatter setup' do
+        config.reporter
+
+        expect(config.formatters).to be_empty
+      end
+
+      it 'buffers deprecations until the reporter is ready' do
+        allow(config.formatter_loader).to receive(:prepare_default).and_wrap_original do |original, *args|
+          config.reporter.deprecation :message => 'Test deprecation'
+          original.call(*args)
+        end
+        expect {
+          config.reporter.notify :deprecation_summary, Notifications::NullNotification
+        }.to change { config.deprecation_stream.string }.to include 'Test deprecation'
+      end
+
+      it 'allows registering listeners without doubling up formatters' do
+        config.reporter.register_listener double(:message => nil), :message
+
+        expect {
+          config.formatter = :documentation
+        }.to change { config.formatters.size }.from(0).to(1)
+
+        # notify triggers the formatter setup, there are two due to the already configured
+        # documentation formatter and deprecation formatter
+        expect {
+          config.reporter.notify :message, double(:message => 'Triggers formatter setup')
+        }.to change { config.formatters.size }.from(1).to(2)
+      end
+
+      it 'still configures a default formatter when none specified' do
+        config.reporter.register_listener double(:message => nil), :message
+
+        # notify triggers the formatter setup, there are two due to the default
+        # (progress) and deprecation formatter
+        expect {
+          config.reporter.notify :message, double(:message => 'Triggers formatter setup')
+        }.to change { config.formatters.size }.from(0).to(2)
+      end
+    end
+
     describe "#default_formatter" do
       it 'defaults to `progress`' do
         expect(config.default_formatter).to eq('progress')
@@ -1480,8 +1527,13 @@ module RSpec::Core
       end
 
       context 'when no other formatter has been set' do
+        before do
+          config.output_stream = StringIO.new
+        end
+
         it 'gets used' do
           config.default_formatter = 'doc'
+          config.reporter.notify :message, double(:message => 'Triggers formatter setup')
 
           expect(used_formatters).not_to include(an_instance_of Formatters::ProgressFormatter)
           expect(used_formatters).to include(an_instance_of Formatters::DocumentationFormatter)